+++ HISTORY.md
@@ -8,14 +8,17 @@
 - [#265] - Respect configuration from `.npmrc`. ([@zkochan])
 - [#6] - Use junctions on Windows. ([#269], [@zkochan])
 - [#270] - Support basic authorization. ([#278], [@leipert])
+- [#144] - Uninstall support. ([#268], [@zkochan])
 
 [#6]: https://github.com/rstacruz/pnpm/issues/6
 [#21]: https://github.com/rstacruz/pnpm/issues/21
+[#144]: https://github.com/rstacruz/pnpm/issues/144
 [#257]: https://github.com/rstacruz/pnpm/issues/257
 [#258]: https://github.com/rstacruz/pnpm/issues/258
 [#259]: https://github.com/rstacruz/pnpm/issues/259
 [#262]: https://github.com/rstacruz/pnpm/issues/262
 [#265]: https://github.com/rstacruz/pnpm/issues/265
+[#268]: https://github.com/rstacruz/pnpm/issues/268
 [#269]: https://github.com/rstacruz/pnpm/issues/269
 [#270]: https://github.com/rstacruz/pnpm/issues/270
 [#278]: https://github.com/rstacruz/pnpm/pull/278